+++ src/MultipartStream.php
@@ -75,7 +75,7 @@ final class MultipartStream implements StreamInterface
             $str .= "{$key}: {$value}\r\n";
         }
 
-        return "--{$this->boundary}\r\n".trim($str)."\r\n\r\n";
+        return "--{$this->boundary}\r\n".rtrim($str, "\r\n")."\r\n\r\n";
     }
 
     /**